(feature) Allow basic auth to be enabled throughout the app
So that we can deploy our changes to a staging server before we are
ready to give access to the general public, we need to protect the site
from unwanted access.

To keep things simple, I've added HTTP Basic Authentication. This is
only enabled once the `HTTP_AUTH_USER` and `HTTP_AUTH_PASSWORD`
environment variables are set.

I've wrapped this functionality into a concern, so it can be tested in
isolation.