(feature) Delegate fetching properties to an API
This forces us to drive out how we're sending the search data from the
form to the API.

- For now let's treat HackneyAPI as the boundary of the system and have
it be a class that has methods which return JSON (hence the string
keys). Later we might push this down further to a base JSON API (which
will probably just be Faraday)
- The AddressSearch code is maybe YAGNI at the moment (we could just
call 'postcode' on the form object), but I'm keen to separate out early
the difference between the data which we _validate_ (and potentially
show back to the user) from the data which we _sanitise_ (and send to
the query) since we've been bitten by this in the past
- I decided to name the result objects 'Property': in this context we're
using the addresses, but the data we're getting is property data.
- Note the use of 'spy' in one of the tests: We haven't used this much
but I like it for this sort of assertion: https://relishapp.com/rspec/rspec-mocks/docs/basics/spies