(feature) SelectedAnswerStore stores answers in the session
This approach of storing answers in the session was taken directly from
https://github.com/alphagov/verify-frontend. It provides a lightweight
wrapper around Rails' session store.

Pass a hash of values for each page - probably corresponding to the form
fields on that page.

Note: because session stores can't in general handle symbols, the data
can only be accessed using string keys