(feature) Pass property_reference around, but show full address
The radio button on the address search results page passes the
property_reference around, but we need to show the full address
to the user, and also store this information in the session.

In order to do this, I've expanded the JsonApi endpoint to fake out
two separate API routes. This allows different results to be
returned for requests to get properties by postcode vs. get details
of an individual property.

- Using the SelectedAnswerStore to store answers and use that to fetch
the individual property from the API