(feature): normalise postcodes in form output
Ensure postcodes are in a standard format before we send them to the
query object.

Is this YAGNI?

Should this be the responsibility of a different object?

I'm conflicted on whether this is better to do only on output, or when
presenting back to the user as well. If we did it up front we'd have to
override the getter or setter, and we'd have to handle nil.