(feature) Addresses controller handles nil refs
If the property_reference is nil, then something has gone quite wrong -
someone has posted to this action with invalid parameters. Still, best
to catch that rather than let it go through to the API call.

This doesn't have a test - I don't think an integration spec is
justifiable (?). To unit test this we would extract a form object, but
we agreed not to do that for now.