(chore) Use SimpleForm's 'input' to render radio buttons
Replace SimpleForm's `collection_radio_buttons` with the more high-level
`input` method which adds additional class names to the resulting HTML
markup, so that it can be styled correctly.

Additionally, add the 'button' class on the submit button.

NB: The GOVUK radio buttons don't work properly with
Capybara/Poltergeist tests because the styling hides the real form
control. Capybara can't click on hidden elements. Because of this, I've
changed the approach to click the associated label element instead.

To DRY this up, I created a capybara helper module which is included in
all feature specs.