(fix) Backfill stubbing in feature spec
I think this was working because the JsonApi class was returning a
hardcoded response, and this test (correctly) doesn't look at any data
in the API response.

However, we now want to replace that hardcoded response with a real
implementation, so we need to explicitly stub out these calls to the API

The simplest way to make this test work is to have it return an empty
list of results in both cases.