(feature) Confirmation page (dummy)
This is mostly hardcoded - except for the repair reference.

Mocked up a Confirmation object as a Struct because I don't want to be
passing 20 individual instance variables to the view ;)

The spec here was just to drive out the routing and controller. When we
have real templates for everything, I anticipate this being a full
end-to-end test which goes through all the screens and checks that the
correct values are shown on the confirmations page.

Need to remember to pull out the phone number to i18n at some point

NOTE: Repair references look like they're probably sequential - this is
a security issue.