(feature) Residents can navigate back in the form
Go through each page as simply as possible and then click on all the
back links all the way back to the start page

The pending specs are there to remind us to implement the back links on
the less obvious pages.

Diagnosis pages are not implemented yet, so skip that page for now.

In this initial implementation we hardcode the links into the page -
we'll make that more flexible in the next step.