(feature) Presenter class for 'back' links
These links are going to be on multiple pages so we need a consistent
way of generating them.

This approach assumes that we'll always be redirecting back to the index
action of some controller

The test is a bit fiddly since basically we're just composing bits from
I18n and the route helpers. We'd get that for free if we did this as a
helper, but then we'd have to pass some information to the view which
the view probably shouldn't know, so I prefer to have this in an object
which can be built in the controller.