(feature) Contact details can be persisted
A class with the responsibilities for 1. persisting the answers and 2.
knowing what data to put there. If we decide that this should be rolled
into the form objects then that should be fairly easy, but I like how
this way we get to keep a (tested) "valid?" method on the form and use
it in the "save" method here.

Maybe at some point the 'persist_answers' method gets extracted to its
own object, since the rest of the code is generic, but that feels like a
step too far at this stage.