(chore) Start form shows error when submitted without answer
Previously, I used a custom validator which set the error message on
`@form.errors[:base]`, which feels odd as the other forms use standard
attribute validations.

To resolve this, I've renamed the sole form attribute to `answer`, which
means Rails automatically provides "Answer cannot be blank" by default.

It also has the benefit of showing the error inline, with the field
highlighted correctly and the correct ARIA attributes applied.