(chore) User gets an error if they don't select an address
If the user doesn't select an address, show them an error and redisplay
the address search results.

To achieve this, first the entered postcode is stored in a hidden field
so it can be reused if an error occurs.

The purpose of the `Address` model wasn't particularly clear, so I've
renamed it to `AddressForm`. This change means that the validation of
the required field `property_preference` can be moved into it.

Additionally, when re-rendering the search results within
`AddressesController`, we can use the form object rather than
interrogating `params` directly.