(feature) Callback time in confirmation body
It was previously only in the summary.

Changed the wording so that it makes sense in both contexts.

Using 'in the morning (8am - 12pm)' would be tricky, because of the all
day case: 'in the all day...' doesn't work, and 'Callback time: in the
morning ...' feels odd as well.

Decided against moving to locales at the moment: feels fiddly and we
have bigger fish to fry. Main thing for me right now is that the strings
aren't hardcoded into the view.