(feature) Repair created at the end of the form
Note that this pushes the controller over the ABC complexity limit.

The CreateRepair service is mainly responsible for translating the
selected answers into params that can be sent in the repair creation
request.

The CreateRepair::Params object is maybe premature, but I can see us
wanting to extract private methods when we're sending many more
parameters.