(refactor) move gas notice above start button
Based on user feedback: They didn't see the gas message when it was
below the Start button. The idea here is that by putting it above, but
not making it more prominent then it will get read without detracting
from the main call-to-action.

Changed from a %div to a %p to ensure that there's the correct amount of
padding between this notice and the button.

Removed 'aside' since I'm not sure we're thinking of this as an aside
anymore.