(feature) Form submit buttons are always shown
We thought this was a GDS-ish pattern, but actually it's not, and leads
to complications e.g. when revisiting a form (an option is already
selected, but the button is not visible)

The purpose of the deleted test was to test the showing/hiding of the
submit button, so it's no longer relevant.