(feature) Fetch data for an answer from a question
This will be needed by other classes which need to know what to do based
on an answer.

Could have just returned nil or an empty hash when the key doesn't
match - not sure what the most appropriate response is, but at least
with an exception we'll know if something goes wrong.