(fix) Lazy-evaluate fallback request reference key
This caused a spec to fail.

The problem was that even if 'requestReference' was present then
@result.fetch('repairRequestReference') was being evaluated, and failing
with a missing key error.

Passing a block is a simple way to ensure that it's only evaluated if
the first part fails. It's ugly, but we'll be able to remove this when
the repairs API is fixed.