(refactor/spec) Move 'diagnosed repair' feature
Since we'll now need to have a new confirmation page for diagnosed
repairs, it makes more sense to have this alongside the other main
end-to-end feature spec.

I've moved it so that it currently passes, and in the next commit we'll
make the required changes for this feature - hopefully that will make it
easier to see what's different.