(refactor) "ContactDetails" => "...WithCallback"
We're going to need two different contact details pages: one with an
callback time section, the other without. We could manage this with
conditionals in the view and controller, but I feel like it will work
out a lot simpler if we have these as separate controllers which refer
to some shared objects.

The one thing I didn't rename was the form fields in the page: these are
still "contact_details", not "contact_details_with_callback" - not sure
if this is justified. Happy to change this if anyone can think of a good
reason to.