(refactor) Extract a ContactDetailsForm
...and define ContactDetailsWithCallbackForm in terms of it.

I'd normally prefer to use Decoration rather than inheritance here, but
with ActiveModel objects that becomes fiddly where errors are concerned:

Because we will want to call e.g. `form.errors`, the errors need to be all on the
same object, but when using decoration you would e.g. have the contact
details errors on the decorated ContactDetailsForm, and only the
callback time errors will be on the ContactDetailsWithCallbackForm.

To get around that you end up having to get any errors on the decorated
object and reapply them onto the decorator. This quickly gets
complicated.

As long as we don't end up needing
ContactDetailsWithCallbackAndSomethingElseForm then hopefully
inheritance should be fine for us.