(refactor) Make RepairParams a first-class citizen
We need to ask the question "is there an SOR code?" in several different
places, and there should be one way of doing that.

I've left the specs of CreateRepair as they are: they act as integration
tests of this class. I'd be happy to remove or simplify some of them
though.