(refactor) Extract a Callback presenter
We're going to want to show either a Callback or an Appointment on the
confirmations page, most likely not both, and both of them fulfil the
same kind of function on this page: they are both an 'action' which has
been scheduled.

Rails supports passing renderable objects to `render`, as long as they
have a method saying where to find the corresponding partial. We can
exploit this by setting an object in the Confirmation presenter, meaning
that the controller and view don't need to know that there's a
difference at all.

Move the presentation of the time slot into the Callback method. I
think it's OK for the Confirmation class to not care too much about the
scheduled_action objects: they can be unit tested and are integration
tested through the feature specs.