(refactor/spec) Helper function: stub diagnosis
There are a lot of test cases where we're stubbing out the diagnosis
questions - in most cases we want to stub one question.

Moving this into a function and out of the 'before' removes the 'mystery
guest' problem (i.e. "where is that 'skip' coming from"), while keeping
the stubbing to one line in most cases.

https://robots.thoughtbot.com/mystery-guest

Perhaps this method goes too far in the case where an ID and question
title are specified - I'd be happy to reverse this change for those
specs.