(refactor) AddressSearch facade
The Addresses controller is getting too long - we need to extract
something to deal with that.

This is the 'pass one object to the view' approach from Thoughtbot in
response to Sandi Metz's rules:

https://robots.thoughtbot.com/sandi-metz-rules-for-developers#only-instantiate-one-object-in-the-controller

One advantage is that we initialise all the objects required by the page
in one place, so we can create the same object when a validation error
occurs and we don't have to worry about keeping things in sync.

The downside is that it becomes a bit more obtuse: the objects passed to
the view are one step removed than they usually are.