(refactor) Repair data can be accessed by object
We're going to need this in a couple of places - specifically to display
the work order reference in the confirmation page.

At the moment this is just dealing with access to a hash, but the
structure of this hash is going to change shortly, so this class is a
way of insulating other code from that change.