(fix) Confirmation page uses repair request id
Previously we were passing either a work order reference or repair
reference to the confirmation page depending on where we're coming from.
This is clearly wrong: when we want to fetch the confirmation data from
the repair request we'll want to be fetching the repair request
records, so that's the ID we need.

In order to get the Work Order reference we need to request the created
repair again.

In the confirmation specs: if we don't pass an explicit value for `api:`
then the tests will attempt to actually load `HackneyApi`, but that
isn't defined in the context of these tests.

It's irritating to have to add `api: double` everywhere, but I guess in
a way this is testing that those methods don't interact with the API at
all.