(refactor) Use javascript for back links
Our previous approach (the controller defines where the back link goes)
was fragile, incomplete and complicated. Instead, we have decided to
follow the approach of some Gov.uk pages and use a Javascript-powered
link which mirrors the behaviour of the browser's back button.

In the event of Javascript not being enabled, no links will show.

Note that this link does say that links should work without Javascript,
but we haven't found a reasonable way to make that happen:
https://www.gov.uk/service-manual/design/question-pages#back-link

- Remove the `Back` object
- when javascript is enabled, inject a javascript back button into the
  page
- Use the 'link-back' class to style the link with a triangle (this
  style comes from the gds toolkit)
- Add back links to the question pages
- Update the tests to reflect that:
  - when you go back from a validation error, you go back to the form
    without a validation error
  - when you go back from the address selection page, you now go back to
    the address search page (which is what the browser back button does.