(feature) Allow user to pick from a list of available appointments
The routes in the Rails app for appointments now reference the
`repair_request_reference` rather than the `work_order_reference`. This
is because we can always use the API to fetch the latter with the
former, but not the other way around.

I've reused the `Repair` class to extract the references from the
response from the API.