(refactor) SelectedAnswerStore into app controller
This is used in almost every controller

This is the approach used in the gov.uk Verify app:

https://github.com/alphagov/verify-frontend/blob/verify_frontend_release_342/app/controllers/application_controller.rb#L90-L92

Memoised so that it's only loaded once per request. This probably means
that controller actions can't set an answer, and read out that answer in
the same request, but that would probably not be a good use of
SelectedAnswerStore anyway.