(feature) special describe pages use partials
At the end of some of the diagnosis paths we want to have some specific
content on the required 'describe repair' page so that we can get the
resident to provide information which will be useful in diagnosing their
problem.

Previously our approach only allowed for changing the title. Now we
allow arbitrary html by using partials, specified with a ':desc' key in
the questions.yml