(feature) Reject partials which don't exist
It's tenuous, but we don't want an attacker to be able to render
arbitrary partials in the page. We can use a method provided by Rails to
check whether a partial exists in a particular directory.

DescriptionPartialChecker is a thin wrapper around lookup_context to
abstract away it's awkward signature. That third boolean argument
determines whether the method searches for partials or non-partials.

Note that we can't pass DescriptionPartialChecker in as a default
argument to DescribeRepair because lookup_context is initialized by
Rails (it contains all the template paths).