(chore) set a minimum coverage
This PR was the final piece in completing our test coverage: the only
remaining gaps are on files generated by Rails which we're not using
yet. It would be good to encourage this to stay high.

This setting will make RSpec return a failure status code if the code
coverage goes below 98%, with a simplecov message explaining why. This
won't block anything locally but will cause the RSpec checks to fail on
CI if code coverage goes down.

I think it's fine for someone in future to push this number down, but
that should be a deliberate choice, rather than as a result of
forgetting to cover part of the code.