(feature) API requests are made with client certs
- Extract a class for building the connection so that we can extract
  some private methods. This might be something we decide to inject
  later, but I'm not sure it makes sense, since JsonApi methods are
  rescuing Faraday errors, so there's inherent coupling there
- Rescue and re-raise the case where the private key is nil, because
  otherwise the exception is "TypeError: can't convert nil to String".
  Otherwise the Key and Cert errors seem useful as they are.
- Update README with Proxy SSL variables
- Inherit specific errors from the generic error class, so that callers
  can rescue from all errors in this class by rescuing  JsonApi::Error

I couldn't work out how to test that the cert is used if provided. The
approach of these tests so far has been to avoid referring to Faraday at
all and rely on webmock. I think the correct approach might be for the
webmock response to be a request for a certificate, but I can't work out
how to do that.

Sadly it may not be worth testing this: if it doesn't work then none of
the API calls will work, so we'll find out pretty quickly!

We at least have a test which ensures that when we pass a valid cert and
key that a request is made.