(refactor) Extract a QuestionSetLoader
Adding a new parameter to QuestionsValidator makes the QuestionSet
intializer too long, so this feels like a good opportunity to split out
a loader class, which is a step in the direction of loading questions
once and accessing them in many places.

Note that the fact that `questions` are passed as an initialization
parameter to QuestionsValidator prevents us from injecting that class as
well. Not too bothered by that: we can do that refactoring in the future
if it seems useful.