(feature) Questions have IDs
Previously the ID was just the key used to refer to a question. Now we
need to know what the ID of the question was, and passing it as part of
the Question object feels better than passing that as an additional
parameter from the controller.

Unfortunately this requires changing a lot of specs...