(fix) Don't display room for communal areas
We didn't have any specs for non-internal spaces. Not sure if this one
is good enough.

I don't like having a conditional in the view. Perhaps it would be
better to have a single 'location' field which either has the room,
Outside or Communal. That would get rid of the awkward repeated 'room'
key. Or maybe this is OK for now?