(refactor) Extract common rescues in JsonApi
This list of rescues is only going to grow so best to extract it now.

This approach is inspired by a StackOverflow post:
https://stackoverflow.com/a/31387752/1035431

...and an Avdi Grimm talk:
https://youtu.be/T8J0j2xJFgQ?t=24m49s

I don't like the nested block here, but the alternative would be to have
separate methods e.g. `basic_get`\`basic_post`, and I'm not convinced
that that would be any better.

I've separated out the call to `.body` to make get and post more
symmetrical