(refactor) Pass question Id in as part of the hash
This does feel slightly better (fewer parameters is usually better), but I don't
think we get away without updating the tests because otherwise we're
using unrealistic test data.

I think the solution to DRYing this up would be to introduce a factory
for creating questions (I don't mean FactoryBot, just a helper method),
but I'd probably only want to use that for the feature specs anyway.