(feature) Session is reset before reporting a new repair
Reset the session when starting to work through the service a second
time. This ensures that users don't see previous answers in the forms
and also prevents a situation where the repair has both a `sor_code` and
a `callback_time`, which should be mutally exclusive.

The feature spec runs through the process from beginning to end twice:

 - The first time, we don't reach diagnosis, and prompt the user for a
   callback time
 - The second time, we are able to diagnose, so a SOR code is chosen and
   an appointment booked

Before resetting the session, the confirmation page would incorrectly
show the callback message. Now it show the appointment time as expected.