(chore) Add 'rel="nofollow"' to back links
Originally, I was going to add `rel="nofollow"` to the submit buttons as
well, as requested. However, `rel` is only a valid attribute for `a`,
`area` and `link` tags, according to the HTML5 specs.

Additionally, Google won't visit pages only accessible via a POST, to
prevent it from accidentally submitting random data to forms all over
the web.