(feature) Include contact name and phone number when creating a repair
Now the `POST /v1/repairs/` endpoint accepts a hash of contact details,
we should include them!

Additionally, I've had the change `ContactDetailsSaver` to use strings
for its keys rather than symbols. We're using a combination of the two
in various places, and in this case it was preventing the contact name
and phone number being read of out of the answer store.