(feature) Include callback time in description
Format it in the same way as it is displayed on the confirmation page.

Includes fix:
Callback time gets saved as a symbol, and read out as a
symbol (we're expecting a string key). This is because it hasn't been
saved to the session cookie and read out again (that's what casts the
symbols to strings).

We should probably write everything as strings, since the casting of
symbols to strings is a little confusing, but that should happen
separately to this PR.