(chore) workOrderReference is nested inside workOrders
Previously, `orderReference` was returned at the root level of the JSON
response. Now, it is correctly nested inside `workOrders`. This is so
the endpoint can be used for retrieving repairs that have multiple
attached work orders.

For now, none of the repairs raised through our service will have more
than one work order, so I'm returning the `workOrderReference`
associated with the first work order in the returned array.

I've also added a `require` into `spec/presenters/confirmation_spec.rb`.
Before I added this, this spec file couldn't be run in isolation, as it
complained that `ActiveSupport::Inflector` wasn't loaded.