(refactor/fix) Rationalise all phone numbers
There are a lot of pages in the site which include phone numbers. In
some cases the numbers were inconsistent, and how they were presented on
the page was inconsistent as well.

Now there is only one place where phone numbers are explicitly written,
and everywhere they are used in the pages they are used in the same way:
as a telephone number link with the appropriate 'itemprop'.

Note that in some places the specific number used on a page has been
corrected to the most appropriate one.