(feature) Prevent appointment data being tampered with
Because appointments returned from the API don't have a simple id to
reference them, I used the `beginDate` and `endDate` concatenated
together.

This comes with a major drawback - we're trusting the contents of these
fields between coming out of the API, being used in the user's browser
and being used to book an appointment. So there's a chance that the user
could manipulate the time range to their advantage.

Using `ActiveSupport::MessageVerifier` I encrypt the appointment hash
and use that as the 'value' when it is used to display the appointment
form. This data is then decrypted and then sent to the API in order to
book the appointment.

Now, if someone attempts to alter the value, an exception is thrown.