(spec) Simplify the test case around resetting the session
Originally, I was going through the entirety of the repairs reporting
process twice in order to check that the correct details are shown to
the user on subsequent visits to the site.

Based on @dgmstuart's suggestion, I'm going up to the point where a
successful diagnosis occurs, and then on the second visit, checking that
the contact details page contains the `callback_time` checkboxes.

As this is only one page away from the confirmation page, I decided to
keep the existing spec that continues through to this page and checks
the returned content.