(feature) Check that all referenced pages exist
We found in testing that the 'blocked_only_toilet' page doesn't exist,
so we need a check for pages to make sure we don't have any others.

The check uses the same approach as the DescriptionPartialChecker. The
third argument to `lookup_context.exists?` defines whether it is a
partial or not, and defaults to false.

Note that normally we could check for pages with:

    Rails.application.routes.recognize_path

However, that doesn't work with HighVoltage, since the page name is an
argument, not part of the route as recognized by Rails. In any case,
recognize_path doesn't seem to be documented so is maybe not public api.