(refactor) Inject questions validator
We were passing the partial_checker down three levels of object, and now
that we want to add another checker (page checker) that seems crazy. The
downside is not having questions as an instance variable in the
validator, but I think that's a small price to pay.

This code now separates the dependencies from the arguments, which feels
like the right direction.