(chore) Outside 'describe repair' text shouldn't mirror question
Previously, if you answered "What is the problem outside your home?"
with "something else", the next page would show the same answers a
second time, only as bullet points on the free-text description page.

I've therefore amended the description page text to be similar to the
`_communal_describe_repair` one - albeit with a tweak to the request
for the location - as it'll be on the resident's property, rather than
elsewhere on the estate.