(fix) Flow through toilet not flushing / blocked
The "unblock your toilet" video page was incorrectly being shown to
users selecting "toilet not flushing". I've changed this to be shown
to users with "blocked toilet" instead.

From that page, if they click "Continue" they are shown a new question
"Is this your only toilet?".

Choosing "yes", shows the emergency contact page. Choosing "no", sets
the SOR code and asks them to provide further details in the normal way.