(feature) Include last question in the description
Identify these as being the ones with a 'desc' key (i.e. it's
explicitly the last question), or if it has no 'next' or 'page' (i.e.
it's implicitly the last question.

The fact that this is implicit feels fragile, but I guess it makes it
easier to write the questions.yml

Made last question mandatory because if it's not there then something
has gone really wrong. Unfortunately this results in a bunch of tests
failing, but I think that's more of a sign of our tests not being
specific enough.