(feature) Leaseholders cannot report repairs
The API 'get property' endpoint will shortly return a new `maintainable`
boolean.

When this is `false`, it is likely the property belongs to a leaseholder
or is a non-Hackney Council managed address.

In this case, redirect the user to a page telling them to call
the Repairs Contact Centre for further assistance.