(refactor) AddressSaver handles address not here
Switch to using a case statement: this controller can redirect in
multiple ways, so let's make it simpler to do that.

Here we extract a method to keep the complexity and method length down
(this lets us finally get the ABC complexity within the limit). I feel
kind of OK about this since the purpose of the method is to select
between url helpers (which feels like a controller responsibility) and
the logic involved hasn't really got anything to do with this particular
controller.