(chore) remove communal/outside repairs
These can't be raised against properties (need to be against blocks or
estates instead), so we can't handle them yet. Instead bump the 'which
room?' question to be the first, and change the wording accordingly.

Hopefully reverting this commit will allow us to reinstate them, but I'm
not 100% sure it'll be that simple.