(fix) YouTube embeds don't overflow screen on mobile
Whilst doing browser testing, we discovered that the changes made to
support the YouTube video embed in IE8 caused them to overflow the edge
of the page on mobile.

Now videos are displayed at the recommended size of 560px wide on
desktop, but restricted to 100% width (if this is less than 560px) on
mobile.