(fix) Boiler issues are an emergency
As it's the winter, heating issues need to be raised with the RCC. This
ensures they are dealt with within 24 working hours.

I've also merged the "Hot water" and "Heating" answers, as they both result
in the same page being shown, which has the options "I have no heating" and
"I have no hot water".