Remove low value tests for problem description
The tests removed from `repair_params_spec.rb` checked that blank
descriptions get correctly shown as "No description given" for various
combinations of parameter. Given the problem, this kind of exhaustive
checking doesn't really add any extra confidence after the first test,
and it makes evolving the behaviour of the system a pain with all the
extra tests that need to be updated.

The test removed from `create_repair_spec.rb` was originally created to
ensure that a completely blank description would be sent as the string
"n/a" because the API required a value. When the behaviour of the system
was changed to always create a meaningful description of some sort, the
test was kept. It's now duplicating the tests done in
`repair_params_spec.rb` at an awkward level of abstraction, causing pain
for refactoring and behaviour changes without providing any extra
confidence.