Redo problem description
- Show metadata above freeform description
- Replace "Last question" with "Problem with"
- Omit "Room: Other" from metadata

From testing with DLO operatives the structured data turns out to be
more consistently useful to them that the free text description from the
resident. Moving the metadata above and compacting should provide a
better experience for them.

Including non-specific metadata like "Room: Other" and talking about
"Last question" could be confusing for the operatives assigned the
repairs unless they know the system that has generated it. As they're
unlikely to have seen nor used the resident facing system, use a clearer
"Problem with" phrasing and only include the room when it's a specific
place (at the moment only Bathroom or Kitchen).