Add repair as button and notice on description page
Resolves trello card
https://trello.com/c/CApWT7jI/64-improve-add-another-repair-design-on-relevant-pages

Made the 'add another repair' link more visible.
Added a notice to the user on repair description page, asking to report
one repair at a time.

For the users wanting to report multiple repairs this would give an
indication that it's possible and doesn't require writing multiple repairs
in the free text area on repair description page.