Filter appointments before 2 business days away
Previously we filtered appointments that were on the same day as the booking: EG if you booked on a Thursday, you could book appointments as early as Friday morning.

This did not give planners enough time to plan accoringly. New rule is to give at least a whole business day before the appointment. EG Booking on Thursday will not allow appointments before Monday morning.