(chore) Creating a repair on the new dashboard requires new estimatedunit value
* This is a new mandatory field in the endpoint for the new API: https://g6bw0g0ojk.execute-api.eu-west-2.amazonaws.com/staging/repairs/swagger/index.html#operations-Repairs-V1RepairsPost
* estimatedunits - reference to this found on Hackney Trello board.
Internally there's a system called Keyfax we use to navigate an officer through to the correct SOR code.
That system CAN (but doen't) produce multiple repairs - which is where the estimatedunits comes in.
But it can be defaulted to "1" as you have here because we don't have that functionality to worry about.