Add DelayedJob
This will allow us to queue background jobs to run asynchronously, rather
than making users wait while we add their repairs to the Google Sheet. I did
consider using Sidekiq for this, as we already have Redis on Heroku, but
using Redis for a cache (as we are for Flipper) and queueing jobs is not
recommended (https://github.com/mperham/sidekiq/wiki/Using-Redis#multiple-redis-instances).